+++ src/aks-preview/azext_aks_preview/tests/latest/test_validators.py
@@ -179,14 +179,17 @@ class TestMaxSurge(unittest.TestCase):
         self.assertTrue('positive' in str(cm.exception), msg=str(cm.exception))
 
 
+class ValidateAddonsNamespace:
+    def __init__(self, addons):
+        self.addons = addons
+
+
 class TestValidateAddons(unittest.TestCase):
     def test_validate_addons(self):
         addons = list(ADDONS)
         if len(addons) > 0:
             first_addon = addons[0]
-            namespace1 = {
-                "addons": first_addon
-            }
+            namespace1 = ValidateAddonsNamespace(first_addon)
 
             try:
                 validators.validate_addons(namespace1)
@@ -195,16 +198,44 @@ class TestValidateAddons(unittest.TestCase):
 
             midlen = int(len(first_addon)/2)
 
-            namespace2 = {
-                "addons": first_addon[:midlen] + first_addon[midlen+1:]
-            }
+            namespace2 = ValidateAddonsNamespace(
+                first_addon[:midlen] + first_addon[midlen+1:])
             self.assertRaises(CLIError, validators.validate_addons, namespace2)
 
-        namespace3 = {
-            "addons": "qfrnmjk"
-        }
+        namespace3 = ValidateAddonsNamespace("qfrnmjk")
 
         self.assertRaises(CLIError, validators.validate_addons, namespace3)
+
+
+class AssignIdentityNamespace:
+    def __init__(self, assign_identity):
+        self.assign_identity = assign_identity
+
+
+class TestAssignIdentity(unittest.TestCase):
+    def test_invalid_identity_id(self):
+        invalid_identity_id = "dummy identity id"
+        namespace = AssignIdentityNamespace(invalid_identity_id)
+        err = ("--assign-identity is not a valid Azure resource ID.")
+
+        with self.assertRaises(CLIError) as cm:
+            validators.validate_assign_identity(namespace)
+        self.assertEqual(str(cm.exception), err)
+
+    def test_valid_identity_id(self):
+        valid_identity_id = "/subscriptions/testid/resourceGroups/MockedResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/mockIdentityID"
+        namespace = AssignIdentityNamespace(valid_identity_id)
+        validators.validate_assign_identity(namespace)
+
+    def test_none_identity_id(self):
+        none_identity_id = None
+        namespace = AssignIdentityNamespace(none_identity_id)
+        validators.validate_assign_identity(namespace)
+
+    def test_empty_identity_id(self):
+        empty_identity_id = ""
+        namespace = AssignIdentityNamespace(empty_identity_id)
+        validators.validate_assign_identity(namespace)
 
 
 if __name__ == "__main__":