+++ src/virtual-wan/azext_vwan/commands.py
@@ -93,12 +93,13 @@ def load_command_table(self, _):
         g.custom_command('list', 'list_virtual_hubs')
         g.generic_update_command('update', custom_func_name='update_virtual_hub', setter_arg_name='virtual_hub_parameters', supports_no_wait=True)
 
-    with self.command_group('network vhub connection', network_vhub_sdk) as g:
+    with self.command_group('network vhub connection', network_vhub_sdk, resource_type=CUSTOM_VHUB_ROUTE_TABLE) as g:
         g.custom_command('create', 'create_hub_vnet_connection', supports_no_wait=True)
+        g.wait_command('wait')
 
     resource = 'virtual_hubs'
     prop = 'virtual_network_connections'
-    with self.command_group('network vhub connection', network_util) as g:
+    with self.command_group('network vhub connection', network_util, resource_type=CUSTOM_VHUB_ROUTE_TABLE) as g:
         g.command('delete', delete_network_resource_property_entry(resource, prop))
         g.command('list', list_network_resource_property(resource, prop))
         g.show_command('show', get_network_resource_property_entry(resource, prop))
@@ -130,12 +131,13 @@ def load_command_table(self, _):
         g.show_command('show')
         g.generic_update_command('update', custom_func_name='update_vpn_gateway', supports_no_wait=True, setter_arg_name='vpn_gateway_parameters')
 
-    with self.command_group('network vpn-gateway connection', network_vpn_gateway_sdk) as g:
+    with self.command_group('network vpn-gateway connection', network_vpn_gateway_sdk, resource_type=CUSTOM_VHUB_ROUTE_TABLE) as g:
         g.custom_command('create', 'create_vpn_gateway_connection', supports_no_wait=True)
+        g.wait_command('wait')
 
     resource = 'vpn_gateways'
     prop = 'connections'
-    with self.command_group('network vpn-gateway connection', network_util) as g:
+    with self.command_group('network vpn-gateway connection', network_util, resource_type=CUSTOM_VHUB_ROUTE_TABLE) as g:
         g.command('delete', delete_network_resource_property_entry(resource, prop))
         g.command('list', list_network_resource_property(resource, prop))
         g.show_command('show', get_network_resource_property_entry(resource, prop))
@@ -182,4 +184,10 @@ def load_command_table(self, _):
         g.show_command('show')
         g.generic_update_command('update', custom_func_name='update_p2s_vpn_gateway', supports_no_wait=True, setter_arg_name='p2_svpn_gateway_parameters')
         g.wait_command('wait')
+
+    resource = 'p2s_vpn_gateways'
+    prop = 'p2_sconnection_configurations'
+    with self.command_group('network p2s-vpn-gateway connection', network_util, resource_type=CUSTOM_VHUB_ROUTE_TABLE, min_api='2020-04-01', is_preview=True) as g:
+        g.command('list', list_network_resource_property(resource, prop))
+        g.show_command('show', get_network_resource_property_entry(resource, prop))
     # endregion