+++ src/virtual-wan/azext_vwan/custom.py
@@ -218,8 +218,8 @@ def remove_hub_route(cmd, resource_group_name, virtual_hub_name, index, no_wait=
 
 # pylint: disable=inconsistent-return-statements
 def create_vhub_route_table(cmd, resource_group_name, virtual_hub_name, route_table_name, destination_type=None,
-                            destinations=None, next_hop_type=None, next_hops=None, attached_connections=None, next_hop=None,
-                            route_name=None, labels=None, no_wait=False):
+                            destinations=None, next_hop_type=None, next_hops=None, attached_connections=None,
+                            next_hop=None, route_name=None, labels=None, no_wait=False):
     if attached_connections:  # route table v2
         if next_hops is None:
             raise CLIError('Usage error: --next-hops must be provided when --connections is provided.')
@@ -436,10 +436,27 @@ def update_vpn_gateway(instance, cmd, virtual_hub=None, tags=None, scale_unit=No
 def create_vpn_gateway_connection(cmd, resource_group_name, gateway_name, connection_name,
                                   remote_vpn_site, routing_weight=None, protocol_type=None,
                                   connection_bandwidth=None, shared_key=None, enable_bgp=None,
-                                  enable_rate_limiting=None, enable_internet_security=None, no_wait=False):
-    client = network_client_factory(cmd.cli_ctx).vpn_gateways
-    VpnConnection, SubResource = cmd.get_models('VpnConnection', 'SubResource')
+                                  enable_rate_limiting=None, enable_internet_security=None, no_wait=False,
+                                  associated_route_table=None, propagated_route_tables=None, labels=None):
+    client = network_client_route_table_factory(cmd.cli_ctx).vpn_gateways
+    (VpnConnection,
+     SubResource,
+     RoutingConfiguration,
+     PropagatedRouteTable) = cmd.get_models('VpnConnection',
+                                            'SubResource',
+                                            'RoutingConfiguration',
+                                            'PropagatedRouteTable')
     gateway = client.get(resource_group_name, gateway_name)
+
+    propagated_route_tables = PropagatedRouteTable(
+        labels=labels,
+        ids=[SubResource(id=propagated_route_table) for propagated_route_table in propagated_route_tables] if propagated_route_tables else None  # pylint: disable=line-too-long
+    )
+    routing_configuration = RoutingConfiguration(
+        associated_route_table=SubResource(id=associated_route_table) if associated_route_table else None,
+        propagated_route_tables=propagated_route_tables
+    )
+
     conn = VpnConnection(
         name=connection_name,
         remote_vpn_site=SubResource(id=remote_vpn_site),
@@ -449,7 +466,8 @@ def create_vpn_gateway_connection(cmd, resource_group_name, gateway_name, connec
         shared_key=shared_key,
         enable_bgp=enable_bgp,
         enable_rate_limiting=enable_rate_limiting,
-        enable_internet_security=enable_internet_security
+        enable_internet_security=enable_internet_security,
+        routing_configuration=routing_configuration
     )
     _upsert(gateway, 'connections', conn, 'name')
     return sdk_no_wait(no_wait, client.create_or_update,
@@ -719,15 +737,29 @@ def remove_vpn_server_config_ipsec_policy(cmd, resource_group_name, vpn_server_c
 
 def create_p2s_vpn_gateway(cmd, resource_group_name, gateway_name, virtual_hub,
                            scale_unit, location=None, tags=None, p2s_conn_config_name='P2SConnectionConfigDefault',
-                           vpn_server_config=None, address_space=None, no_wait=False):
+                           vpn_server_config=None, address_space=None, associated_route_table=None,
+                           propagated_route_tables=None, labels=None, no_wait=False):
     client = network_client_route_table_factory(cmd.cli_ctx).p2s_vpn_gateways
     (P2SVpnGateway,
      SubResource,
      P2SConnectionConfiguration,
-     AddressSpace) = cmd.get_models('P2SVpnGateway',
-                                    'SubResource',
-                                    'P2SConnectionConfiguration',
-                                    'AddressSpace')
+     AddressSpace,
+     RoutingConfiguration,
+     PropagatedRouteTable) = cmd.get_models('P2SVpnGateway',
+                                            'SubResource',
+                                            'P2SConnectionConfiguration',
+                                            'AddressSpace',
+                                            'RoutingConfiguration',
+                                            'PropagatedRouteTable')
+
+    propagated_route_tables = PropagatedRouteTable(
+        labels=labels,
+        ids=[SubResource(id=propagated_route_table) for propagated_route_table in propagated_route_tables] if propagated_route_tables else None
+    )
+    routing_configuration = RoutingConfiguration(
+        associated_route_table=SubResource(id=associated_route_table) if associated_route_table else None,
+        propagated_route_tables=propagated_route_tables
+    )
     gateway = P2SVpnGateway(
         location=location,
         tags=tags,
@@ -739,16 +771,17 @@ def create_p2s_vpn_gateway(cmd, resource_group_name, gateway_name, virtual_hub,
                 vpn_client_address_pool=AddressSpace(
                     address_prefixes=address_space
                 ),
-                name=p2s_conn_config_name
+                name=p2s_conn_config_name,
+                routing_configuration=routing_configuration
             )
         ]
     )
-    return sdk_no_wait(no_wait, client.create_or_update,
-                       resource_group_name, gateway_name, gateway)
+
+    return sdk_no_wait(no_wait, client.create_or_update, resource_group_name, gateway_name, gateway)
 
 
 def update_p2s_vpn_gateway(instance, cmd, tags=None, scale_unit=None,
-                           vpn_server_config=None, address_space=None, p2s_conn_config_name=None,):
+                           vpn_server_config=None, address_space=None, p2s_conn_config_name=None):
     (SubResource,
      P2SConnectionConfiguration,
      AddressSpace) = cmd.get_models('SubResource',